requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
curlify==2.2.1
facebook-business==16.0.2
frozenlist==1.3.3
httplib2==0.22.0
idna==3.4
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
pycountry==22.3.5
pyparsing==3.0.9
python-dotenv==1.0.0
python-instagram==1.3.2
requests==2.28.2
simplejson==3.19.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
